Fix timeout in regex source generated test
The test is using Roslyn to dynamically emit a Regex with the source generator, as we do throughout the test suite.  However, this test is about validating the default static regex timeout, and the Roslyn code itself uses a regex.  To work around this, this PR splits off the source generated test into its own dedicated one that doesn't rely on Roslyn.