Fix illink tests
The analyzer tests download a ref package, but its version must at least roughly match the version used to run the tests.
Since this updates the SDK to 8.0, update the tests to pick some 8.0 ref package.

Also the new compiler produces two new warnings in one case, so baselined those in the test.