Changed handle table scanning so that we would scan all slots for any heap count we may have.
This necessitated adding a thread_count field to the scan context so the handle table scanning would be aware of the current number of gc threads.

I also did an unrelated small cleanup where I changed the argument list of a few Ref_xxx function to have a ScanContext * instead of an uintptr_t, to reduce the amount of back-and-forth casting required.