PR Feedback and fix failures
The failures are due to Canon instantiations which are not expected in some of the reflection specific parts of the code. Since the change calls the GetDependenciesDueToAccess on new things, it hits this as well. Fixed by checking for Canon instantiations before doing reflection stuff - we don't want the reflection to hit the Canon anyway as those types are not directly reflectable in any form.