JIT: Generalize last-use copy elision for implicit byrefs
Allow omitting the copy if we have _any_ last use before the call, not
just if the candidate itself is dead. For example, for something like

M(s, s.A);

the first argument can now have its copy elided if the access in the
second argument is the last use.

Fix #80494