Fix treatment of fields of implicit byref params with undone promotion
We can get here for locals representing fields of promoted implicit
byrefs that had promotion undone. In those cases we would sometimes
still propagate the last-use information. That's ok in the context of
the optimization we do with the information but it is not technically
correct since another future field access would also be transformed to a
use of the parent local.