Fix backtrace in VS when calling RaiseFailFastException
The memory in the CONTEXT structure was not being full initialized.
Switching from using the transition frame to unwinding filled the stack
with 0xDD values. When VS displayed the backtrace after calling
RaiseFailFastException, it was corrupted (showing a selection of unrelated
frames).

This can reproduced on the main branch by changing the memset here to set
0xDD. So I think this a pre-existing problem.