PR feedback
The biggest change is to use lock free hashtables instead of a dictionary to avoid races.

Note on that - had to avoid the typical GetOrCreateValue because of possible recursion. And also we don't want to store empty records, it's actually very fast to figure out that an entity doesn't have suppressions from the type system, no point in wasting memory to remember that result.aot-run