Avoid caching suppressions to simplify code
After discussing this we think it's not worth the code complexity to cache the suppressions. So this version will always do all of the attribute lookups and parsing on each warning.

Note that it still needs one tiny cache - a hashtable of assemblies, to avoid reporting the warnings produced by suppression parsing more than once.

The code is a bit more complex since I tried to avoid allocating things unnecessarily - there's probably more which can be done in this regard though.