JIT: add concept of edge likelihood
Now that FlowEdges are created early and persist, decorate them with likelihood
information early, if we have edge-based PGO data.

We don't use the likelihood for anything yet, but I need to get it in circulation
so I can start working on refining both initial and subsequent consistency of
the data.

Also add a diagnostic checker for likelhood, and a way to enable it. All of
this is off by default.