disable the TryGetHashCode fast path helper
It was actually making things slower on the interpreter.

In the JIT it made object.GetHashCode about 50% faster for an object
with a pre-computed hash, but that might mean we need the same kind of
"no-wrapper; then wrapper" fastpath/slowpath that we have for the
Monitor.Enter intrinsic in mini.