add CMPXCHG opcodes to the jiterpreter
For the i64 version, pass the address of the i64 values to the helper
function.

The issue is that since JS doesn't have i64, EMSCRIPTEN_KEEPALIVE
messes with the function's signature and we get import errors in the
JITed wasm module.  Passing by address works around the issue since
addresses are i32.