Manual .NET 7 backport of #80218 fixing #79327 (#80881)
Description:

This change puts HFA calculation in Crossgen2 in sync with native
CoreCLR runtime for value types with explicit layout. Previously
Crossgen2 had a shortcut in the routine deciding that structs with
explicit layouts are never marked as HFA that disagreed with
the CoreCLR runtime; consequently, on arm64, Crossgen2 disagreed
with the runtime on whether or not a function returning such a type
should allocate the stack slot for return value, basically
messing up the calling convention and GC refmap, resulting in various
random AVs and corruptions. This was first observed by an internal
customer in WPF apps where MilRectD is the type in question, later
JanK filed the issue 79327 for the same problem.

Customer impact:

Random runtime crashes on arm64.

Regression:

Nope, I believe the incomplete implementation was the original one,
this change just "improves it" by putting it in better sync with
the native runtime. I have also added a code comment mentioning
that these two need to be kept in sync.

Risk:

Low - the error in the previous implementation is obvious, R2RDump
and my new runtime diagnostics clearly show the GC refmap mismatch
caused by this problem and its fixing after applying the Crossgen2 fix.

Link to issue:

Link to PR against main:

Publishing impact:

In the particular case of the WPF app the problem was in
the PresentationCore.dll assembly. The assembly (or rather the entire WPF)
need to be recompiled with Crossgen2 with the fix applied for this to
take effect. For now I assume that is an automated part of the servicing
process.

Thanks

Tomas