Fix src/tests/build.sh argument processing (#81836)
The logic to process test filter args isn't working correctly
because the `shift` happens inside a sourced script, and doesn't
actually shift the outer arguments. The right way to do this is
via `__ShiftArgs`.

Without this fix, `tree nativeaot` was working because
`nativeaot` happened to be a supported argument as well. With the
fix, other test subtrees work as expected.